Refactor: Move split_text_into_chunks to utils module
Moved `split_text_into_chunks` from `deep_research_project/core/research_loop.py` to `deep_research_project/core/utils.py` to separate generic utilities from business logic. Updated references in `research_loop.py` and `tests/test_core_components.py`.
